+++ tests/lib/rules/no-multi-comp.js
@@ -236,6 +236,35 @@ ruleTester.run('no-multi-comp', rule, {
         export default MenuList;
       `,
     },
+    {
+      code: `
+        const MenuList = forwardRef(({ onClose, ...props }, ref) => {
+          const onLogout = useCallback(() => {
+            onClose()
+          }, [onClose])
+
+          return (
+            <BlnMenuList ref={ref} {...props}>
+              <BlnMenuItem key="logout" onClick={onLogout}>
+                Logout
+              </BlnMenuItem>
+            </BlnMenuList>
+          )
+        })
+
+        MenuList.displayName = 'MenuList'
+
+        MenuList.propTypes = {
+          onClose: PropTypes.func
+        }
+
+        MenuList.defaultProps = {
+          onClose: () => null
+        }
+
+        export default MenuList
+      `,
+    },
   ]),
 
   invalid: parsers.all([