+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 
 ## Unreleased
 
+### Changed
+
+* [Meta] npmignore markdownlint config ([#3413][] @jorrit)
+
+[#3413]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3413
+
 ## [7.31.7] - 2022.09.05
 
 ### Fixed