+++ CHANGELOG.md
@@ -5,10 +5,14 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 
 ## Unreleased
 
+### Fixed
+* [`no-unknown-property`]: add `viewBox` on `marker` ([#3416][] @ljharb)
+
 ### Changed
 
 * [Meta] npmignore markdownlint config ([#3413][] @jorrit)
 
+[#3416]: https://github.com/jsx-eslint/eslint-plugin-react/issues/3416
 [#3413]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3413
 
 ## [7.31.7] - 2022.09.05